Basic Tinode Push Gateway adapter.
